mini update game_script.js
Turns out the pie chart is animated by default as I was testing on my pc. On the other Hand those animations don't work on my laptop which I haven't figured out yet why that's the case...
I'll forcefully disable the animations for now for consistency reasons and maybe add them back when I have a solution to this.